adds some changes
