+++ src/main/java/com/school/project/FX_Project/GradeTable.java
@@ -1,4 +1,4 @@
-package com.school.project.FXclasses;
+package com.school.project.FX_Project;
 
 public class GradeTable {
     private int studGradeID;
