+++ src/main/java/com/school/project/FX_Project/Teacher.java
@@ -1,4 +1,4 @@
-package com.school.project.FXclasses;
+package com.school.project.FX_Project;
 
 public class Teacher {
     private int teacherID;
